// src/app/swagger/model/Image.ts
/**
 * UPResourceEditor
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * Contact: ostangenberg@universeprojects.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Image {
    name?: string;

    treePath?: string;

    parent?: string;

    tags?: Array<string>;

    blobKey?: string;

    gcsPath?: string;

    gcsUrl?: string;

    imageUrl?: string;

}